Try to fix Travis build.
Since long time Travis constantly fails with the message "The log length has exceeded the limit of 4 MB (this usually means that the test suite is raising the same exception over and over).".

The fix is to download all dependencies quitely ("--quiet" option) and then run the build with everything already in local repository.